requirements.txt
fastapi
pydantic
google-generativeai
uvicorn[standard]
sqlalchemy
python-dotenv
